# apps/backend/usermng-service/src/docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: FinVerse User Management Service API
  version: "1.0.0"
  description: |
    REST API for managing organizations, members, roles, invitations, and users in the FinVerse platform.

    **Authentication:** Auth0 JWT (RS256) bearer tokens required for all endpoints.

    **Response Format:** All responses follow a standardized envelope structure with:
    - `success` (boolean) - Indicates if the operation succeeded
    - `message` (string) - Human-readable description of the result
    - `data` (object/array) - Response payload (on success)
    - `error` (object) - Error details (on failure)
    - `meta` (object) - Metadata including timestamp and requestId for tracing

    **Request Tracing:** Each request can include an `X-Request-ID` header for distributed tracing.
    If not provided, the server generates a unique ID included in the response metadata.

    **Pagination:** List endpoints return paginated data with page, limit, and total count.

    **Error Handling:** Errors include descriptive messages, error codes, and optional field-level details.

  contact:
    name: FinVerse API Support
    email: api-support@finverse.example.com
    url: https://docs.finverse.example.com

  license:
    name: Proprietary
    url: https://finverse.example.com/license
servers:
  - url: https://api.finverse.example.com/api/v1
    description: Production server
  - url: https://staging-api.finverse.example.com/api/v1
    description: Staging server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0-issued access token (RS256 algorithm).

        **Required Header:** `Authorization: Bearer <token>`

        **Example Scopes:**
        - `admin:manage` - Super admin operations
        - `org:read` - Read organization data
        - `org:write` - Create/update organizations
        - `user:read` - Read user data
        - `user:write` - Update user data
        - `invite:manage` - Create/manage invitations
        - `member:manage` - Manage organization members
  parameters:
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
      description: Optional request ID for distributed tracing
    OrgIdPath:
      name: orgId
      in: path
      required: true
      schema:
        type: string
      description: Organization identifier
...
    InvitationIdPath:
      name: invitationId
      in: path
      required: true
      schema:
        type: string
      description: Invitation identifier
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Page number (0-indexed)
    LimitQuery:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Items per page
  schemas:
...
    ApiSuccessResponse:
      type: object
      required: [success, message, data, meta]
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          description: Response payload
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ApiErrorResponse:
      type: object
      required: [success, message, error, meta]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
        error:
          type: object
          required: [code]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_SERVER_ERROR
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ResponseMeta:
      type: object
      required: [timestamp]
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            success: false
            message: "Validation failed"
            error:
              code: "VALIDATION_ERROR"
              details:
                - field: "name"
                  message: "Name is required"
            meta:
              timestamp: "2024-12-17T10:30:00.000Z"
              requestId: "abc123"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            success: false
            message: "Missing or invalid access token"
            error:
              code: "UNAUTHORIZED"
            meta:
              timestamp: "2024-12-17T10:30:00.000Z"
              requestId: "abc123"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            success: false
            message: "Insufficient permissions for this operation"
            error:
              code: "FORBIDDEN"
            meta:
              timestamp: "2024-12-17T10:30:00.000Z"
              requestId: "abc123"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            success: false
            message: "Resource not found"
            error:
              code: "NOT_FOUND"
            meta:
              timestamp: "2024-12-17T10:30:00.000Z"
              requestId: "abc123"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            success: false
            message: "An unexpected error occurred"
            error:
              code: "INTERNAL_SERVER_ERROR"
            meta:
              timestamp: "2024-12-17T10:30:00.000Z"
              requestId: "abc123"
paths:
...
  /orgs/{orgId}/members:
    parameters:
      - $ref: "#/components/parameters/OrgIdPath"
      - $ref: "#/components/parameters/RequestIdHeader"

    get:
      tags: [Members]
      summary: List organization members
      description: Get paginated list of organization members
      operationId: listOrganizationMembers
      security:
        - bearerAuth: [org:read, member:read]
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Members retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessResponse"
              example:
                success: true
                message: "Organization members retrieved successfully"
                data:
                  items:
                    - user_id: "auth0|123"
                      email: "user@example.com"
                      name: "John Doe"
                  pagination:
                    page: 0
                    limit: 50
                    total: 1
                meta:
                  timestamp: "2024-12-17T10:30:00.000Z"
                  requestId: "abc123"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags: [Members]
      summary: Remove members
      description: Remove multiple members from organization
      operationId: deleteOrganizationMembers
      security:
        - bearerAuth: [org:write, member:manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [members]
              properties:
                members:
                  type: array
                  items:
                    type: string
            example:
              members: ["auth0|123", "auth0|456"]
      responses:
        "200":
          description: Members deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessResponse"
              example:
                success: true
                message: "Organization members deleted successfully"
                data: null
                meta:
                  timestamp: "2024-12-17T10:30:00.000Z"
                  requestId: "abc123"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orgs/{orgId}/invitations:
    parameters:
      - $ref: "#/components/parameters/OrgIdPath"
      - $ref: "#/components/parameters/RequestIdHeader"

    post:
      tags: [Invitations]
      summary: Create invitation
      description: Invite new user to organization via email
      operationId: createOrganizationInvitation
      security:
        - bearerAuth: [invite:manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inviter, invitee, roles]
              properties:
                inviter:
                  type: object
                  required: [name]
                  properties:
                    name:
                      type: string
                invitee:
                  type: object
                  required: [email]
                  properties:
                    email:
                      type: string
                      format: email
                roles:
                  type: array
                  items:
                    type: string
            example:
              inviter:
                name: "Admin User"
              invitee:
                email: "newuser@example.com"
              roles: ["rol_member"]
      responses:
        "201":
          description: Invitation created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessResponse"
              example:
                success: true
                message: "Organization invitation created successfully"
                data:
                  id: "inv_xyz789"
                  organization_id: "org_abc123"
                  invitee:
                    email: "newuser@example.com"
                  invitation_url: "https://finverse.com/accept-invite?id=inv_xyz789"
                meta:
                  timestamp: "2024-12-17T10:30:00.000Z"
                  requestId: "abc123"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      tags: [Invitations]
      summary: List invitations
      description: Get paginated list of pending invitations for organization
      operationId: listOrganizationInvitations
      security:
        - bearerAuth: [invite:manage]
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Invitations retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessResponse"
              example:
                success: true
                message: "Organization invitations retrieved successfully"
                data:
                  items:
                    - id: "inv_xyz789"
                      organization_id: "org_abc123"
                      invitee:
                        email: "newuser@example.com"
                  pagination:
                    page: 0
                    limit: 50
                    total: 1
                meta:
                  timestamp: "2024-12-17T10:30:00.000Z"
                  requestId: "abc123"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /orgs/{orgId}/invitations/{invitationId}:
    parameters:
      - $ref: "#/components/parameters/OrgIdPath"
      - $ref: "#/components/parameters/InvitationIdPath"
      - $ref: "#/components/parameters/RequestIdHeader"

    delete:
      tags: [Invitations]
      summary: Delete invitation
      description: Delete a pending invitation
      operationId: deleteOrganizationInvitation
      security:
        - bearerAuth: [invite:manage]
      responses:
        "200":
          description: Invitation deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessResponse"
              example:
                success: true
                message: "Organization invitation deleted successfully"
                data: null
                meta:
                  timestamp: "2024-12-17T10:30:00.000Z"
                  requestId: "abc123"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...